Add files via upload
SIGNAL GENERATOR



The 4017 IC is a 16-legged IC.



The enable, reset, carry out pins on the 4017 IC are connected to the chassis when using the circuit. After the necessary energy is given to the chassis and pover ends of the integrated and the clock pin in the circuit receives the necessary trigger signal, the integrated outputs are activated at certain intervals. If we look at the circuit, with each triggering signal coming to the 4017 IC, the Q0 output becomes active first from the 4017 outputs and the led connected to this output lights up. Then the other outputs become active respectively and the LED belonging to that output gives light. At the same time, only one output becomes active and the LED of that output lights up and this continues periodically.



The LM555/NE555/SA555 is a highly stable controllercapable of   producing   accurate timing pulses. With amonostable operation, the time delay is controlled by oneexternal resistor and one capacitor. With an astable operation, the frequency and duty cycle are accurately controlled by two external resistors and one capacitor.